// apps/web/src/api/generated/model/approvalAction.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Care API
 * OpenAPI spec version: 0.0.1
 */

export type ApprovalAction = typeof ApprovalAction[keyof typeof ApprovalAction];


export const ApprovalAction = {
  SUBMIT: 'SUBMIT',
  APPROVE: 'APPROVE',
  REJECT: 'REJECT',
  CANCEL: 'CANCEL',
} as const;