// apps/web/src/api/generated/model/jobRole.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Care API
 * OpenAPI spec version: 0.0.1
 */

export type JobRole = typeof JobRole[keyof typeof JobRole];


export const JobRole = {
  DIRECTOR: 'DIRECTOR',
  OFFICE_MANAGER: 'OFFICE_MANAGER',
  SOCIAL_WORKER: 'SOCIAL_WORKER',
  NURSE: 'NURSE',
  NURSE_AIDE: 'NURSE_AIDE',
  PHYSICAL_THERAPIST: 'PHYSICAL_THERAPIST',
  OCCUPATIONAL_THERAPIST: 'OCCUPATIONAL_THERAPIST',
  CAREGIVER: 'CAREGIVER',
  CLERK: 'CLERK',
  DIETITIAN: 'DIETITIAN',
  COOK: 'COOK',
  HYGIENIST: 'HYGIENIST',
  JANITOR: 'JANITOR',
} as const;